unit tests: move fake_path to separate file
