+++ test/unit/vertex_adapter/offset_converter.cpp
@@ -1,65 +1,14 @@
 
 #include "catch.hpp"
+#include "fake_path.hpp"
 
 // mapnik
-#include <mapnik/vertex.hpp>
 #include <mapnik/offset_converter.hpp>
 
 // stl
 #include <iostream>
-#include <vector>
-#include <tuple>
 
 namespace offset_test {
-
-struct fake_path
-{
-    using coord_type = std::tuple<double, double, unsigned>;
-    using cont_type = std::vector<coord_type>;
-    cont_type vertices_;
-    cont_type::iterator itr_;
-
-    fake_path(std::initializer_list<double> l)
-        : fake_path(l.begin(), l.size()) {
-    }
-
-    fake_path(std::vector<double> const &v, bool make_invalid = false)
-        : fake_path(v.begin(), v.size(), make_invalid) {
-    }
-
-    template <typename Itr>
-    fake_path(Itr itr, size_t sz, bool make_invalid = false) {
-        size_t num_coords = sz >> 1;
-        vertices_.reserve(num_coords + (make_invalid ? 1 : 0));
-        if (make_invalid)
-        {
-            vertices_.push_back(std::make_tuple(0,0,mapnik::SEG_END));
-        }
-
-        for (size_t i = 0; i < num_coords; ++i) {
-            double x = *itr++;
-            double y = *itr++;
-            unsigned cmd = (i == 0) ? mapnik::SEG_MOVETO : mapnik::SEG_LINETO;
-            vertices_.push_back(std::make_tuple(x, y, cmd));
-        }
-        itr_ = vertices_.begin();
-    }
-
-    unsigned vertex(double *x, double *y) {
-        if (itr_ == vertices_.end()) {
-            return mapnik::SEG_END;
-        }
-        *x = std::get<0>(*itr_);
-        *y = std::get<1>(*itr_);
-        unsigned cmd = std::get<2>(*itr_);
-        ++itr_;
-        return cmd;
-    }
-
-    void rewind(unsigned) {
-        itr_ = vertices_.begin();
-    }
-};
 
 static double DELTA_BUFF = 0.5;
